Crate relational_graphql

Relational GraphQL is an ORM designed to efficiently translate GraphQL queries into queries against a relational database. It consists of two sections:

A frontend, which most users will interact with, for defining graphql schemas. A graphql schema describes clients’ view of the application’s data model. It provides an ontology that clients can use to conceptualize the various entities and their relationships as well as an expressive language for querying the data.

A backend, which is responsible for translating GraphQL queries into queries against relational databases. The backend implementation is completely agnostic to the specifics of the applications data model. It uses a generic type system describing the GraphQL API in order to automatically compile GraphQL objects and queries into database operations. This makes it possible to write one general backend and reuse it with many GraphQL applications.

This crate comes with a sql backend, which provides a PostgreSQL target as well as a mock database, which is useful for lightweight testing. The library is completely modular, though, so it is possible to implement your own backend to meet your application’s specific needs by implementing the backend traits.

Re-exports

pub use generic_array::typenum;

Modules

graphql

A general interface for defining relational GraphQL applications.

prelude

Common items that you will always want in scope when using GraphQL.

sql

Relational GraphQL backend based on a SQL database (specifically PostgreSQL).

Macros

array

Create an Array from a list of items.

use_backend

Use a certain DataSource implementation as the backend in a GraphQL schema.

Structs

Array

An array of type T with constant length N.

Traits

Length

A convenience for working with [GenericArray].

Functions

init_logging

Initialize tracing.