Module relational_graphql::graphql::backend

Interfaces provided by a backend data source consumed by a GraphQL API.

The entrypoint to this system of traits is DataSource, which describes the interface by which the GraphQL API interacts with the backend data provider. This is, in particular, the glue between the GraphQL and SQL views of the data model, since the SQL model will implement DataSource and the GraphQL layer will interact with the SQL layer exclusively through this trait.

A number of supporting traits are defined here which can be accessed through the DataSource trait by means of its associated types.

Structs

Edge

And edge in a connection, connecting the owner of the connection to another object.

PageRequest

A specification of a page to load in a paginated connection.

Traits

Connection

A Relay-style paginated connection to a collection of objects.

DataSource

A source of data which can be served by the GraphQL API.

Type Definitions

Cursor

An index into a Paginated list of objects.

Paginated

A paginated list of objects.