Trait relational_graphql::graphql::backend::DataSource

pub trait DataSource {
    type Connection<T: Type, C: ObjectType, E: ObjectType>: Connection<C>;
    type Error: Error;

    // Required methods
    fn load_relation<'life0, 'life1, 'async_trait, R>(
        &'life0 self,
        owner: &'life1 R::Owner,
        filter: Option<<R::Target as Type>::Predicate>
    ) -> Pin<Box<dyn Future<Output = Result<Paginated<Self, R::Target>, Self::Error>> + Send + 'async_trait>>
       where R: 'async_trait + Relation,
             Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn load_page<'life0, 'life1, 'async_trait, T, C, E>(
        &'life0 self,
        conn: &'life1 Paginated<Self, T, C, E>,
        page: PageRequest<Cursor<Self, T, C, E>>
    ) -> Pin<Box<dyn Future<Output = Result<Vec<Edge<Cursor<Self, T, C, E>, T, E>>, Self::Error>> + Send + 'async_trait>>
       where T: 'async_trait + Type,
             C: 'async_trait + ObjectType,
             E: 'async_trait + Clone + ObjectType,
             Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn register<'life0, 'async_trait, T>(
        &'life0 mut self
    ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
       where T: 'async_trait + Resource,
             Self: 'async_trait,
             'life0: 'async_trait;
    fn query<'life0, 'async_trait, T>(
        &'life0 self,
        filter: Option<T::ResourcePredicate>
    ) -> Pin<Box<dyn Future<Output = Result<Paginated<Self, T>, Self::Error>> + Send + 'async_trait>>
       where T: 'async_trait + Resource,
             Self: 'async_trait,
             'life0: 'async_trait;
    fn insert<'life0, 'async_trait, T, I>(
        &'life0 mut self,
        inputs: I
    ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
       where I: IntoIterator<Item = T::ResourceInput> + Send + 'async_trait,
             I::IntoIter: Send,
             T: 'async_trait + Resource,
             Self: 'async_trait,
             'life0: 'async_trait;
    fn populate_relation<'life0, 'async_trait, R, I>(
        &'life0 mut self,
        pairs: I
    ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
       where I: IntoIterator<Item = (Id, Id)> + Send + 'async_trait,
             I::IntoIter: Send,
             R: 'async_trait + Relation,
             Self: 'async_trait,
             'life0: 'async_trait;
}

A source of data which can be served by the GraphQL API.

Required Associated Types

type Connection<T: Type, C: ObjectType, E: ObjectType>: Connection<C>

A Relay-style paginated connection to a collection of objects.

THe objects in the collection are of type T. Each object in the collection also represents a relationship, or edge, between the object which owns the collection and the object in the collection. These edges may have additional fields of type E, beyond the fields specified by Relay. The connection itself may also have additional fields of type C, beyond the fields specified by Relay.

type Error: Error

Errors reported while attempting to load data.

Required Methods

fn load_relation<'life0, 'life1, 'async_trait, R>( &'life0 self, owner: &'life1 R::Owner, filter: Option<<R::Target as Type>::Predicate> ) -> Pin<Box<dyn Future<Output = Result<Paginated<Self, R::Target>, Self::Error>> + Send + 'async_trait>>where R: 'async_trait + Relation, Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Load the targets of a Relation.

Errors

Fails if owner does not exist in the data source or if either of the resources related by R have not been registered with register.

fn load_page<'life0, 'life1, 'async_trait, T, C, E>( &'life0 self, conn: &'life1 Paginated<Self, T, C, E>, page: PageRequest<Cursor<Self, T, C, E>> ) -> Pin<Box<dyn Future<Output = Result<Vec<Edge<Cursor<Self, T, C, E>, T, E>>, Self::Error>> + Send + 'async_trait>>where T: 'async_trait + Type, C: 'async_trait + ObjectType, E: 'async_trait + Clone + ObjectType, Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Load a page from a paginated connection.

fn register<'life0, 'async_trait, T>( &'life0 mut self ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>where T: 'async_trait + Resource, Self: 'async_trait, 'life0: 'async_trait,

Register the resource T in the schema.

fn query<'life0, 'async_trait, T>( &'life0 self, filter: Option<T::ResourcePredicate> ) -> Pin<Box<dyn Future<Output = Result<Paginated<Self, T>, Self::Error>> + Send + 'async_trait>>where T: 'async_trait + Resource, Self: 'async_trait, 'life0: 'async_trait,

Get a paginated stream of items matching filter.

Errors

Fails if T has not been registered with register.

fn insert<'life0, 'async_trait, T, I>( &'life0 mut self, inputs: I ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>where I: IntoIterator<Item = T::ResourceInput> + Send + 'async_trait, I::IntoIter: Send, T: 'async_trait + Resource, Self: 'async_trait, 'life0: 'async_trait,

Insert new items into the database.

Errors

Fails if any of the objects in inputs conflict with objects already in the data source (for example, via a duplicate value of a unique field), or if T has not been registered with register.

fn populate_relation<'life0, 'async_trait, R, I>( &'life0 mut self, pairs: I ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>where I: IntoIterator<Item = (Id, Id)> + Send + 'async_trait, I::IntoIter: Send, R: 'async_trait + Relation, Self: 'async_trait, 'life0: 'async_trait,

Insert new pairs into a relation R.

This method relates the objects (owner, target) by R, for each (owner, target) pair in pairs. owner here is the Id of an object in R::Owner, and target is the Id of an object in R::Target.

Errors

Fails if any of the IDs in pairs do not already exist in the data source, or if either of the resources related by R have not been registered with register.