Trait relational_graphql::graphql::OutputType

pub trait OutputType: Send + Sync {
    // Required methods
    fn type_name() -> Cow<'static, str>;
    fn create_type_info(registry: &mut Registry) -> String;
    fn resolve<'life0, 'life1, 'life2, 'life3, 'async_trait>(
        &'life0 self,
        ctx: &'life1 ContextBase<'life2, &'life2 Positioned<SelectionSet>>,
        field: &'life3 Positioned<Field>
    ) -> Pin<Box<dyn Future<Output = Result<ConstValue, ServerError>> + Send + 'async_trait, Global>>
       where 'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait,
             'life3: 'async_trait,
             Self: 'async_trait;

    // Provided methods
    fn qualified_type_name() -> String { ... }
    fn introspection_type_name(&self) -> Cow<'static, str> { ... }
}

Represents a GraphQL output type.

Required Methods

fn type_name() -> Cow<'static, str>

Type the name.

fn create_type_info(registry: &mut Registry) -> String

Create type information in the registry and return qualified typename.

fn resolve<'life0, 'life1, 'life2, 'life3, 'async_trait>( &'life0 self, ctx: &'life1 ContextBase<'life2, &'life2 Positioned<SelectionSet>>, field: &'life3 Positioned<Field> ) -> Pin<Box<dyn Future<Output = Result<ConstValue, ServerError>> + Send + 'async_trait, Global>>where 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait, 'life3: 'async_trait, Self: 'async_trait,

Resolve an output value to async_graphql::Value.

Provided Methods

fn qualified_type_name() -> String

Qualified typename.

fn introspection_type_name(&self) -> Cow<'static, str>

Introspection type name

Is the return value of field __typename, the interface and union should return the current type, and the others return Type::type_name.