Module relational_graphql::graphql::type_system::resource

Resources are complex types in a relational GraphQL API.

A resource type is akin to an object type in GraphQL or a table in a relational database. It has its own fields as well as Relations to other resources. Collections of items of a particular resource can be filter down using a ResourcePredicate. Entire collections of items of a given resource type can also be filtered in or out using a RelationPredicate.

Users can define their own resources by implementing the Resource trait and relatives, or by using the Resource derive macro.

Structs

BelongsTo

A marker type used in the definition of a resource to declare a ManyToOneRelation.

Many

A marker type used in the definition of a resource to declare a ManyToManyRelation.

Traits

Field

Metadata about a field of a resource.

FieldVisitor

Visitor which allows fields of a Resource to describe themselves to a backend.

InputField

A field which must be present when a new object is created.

InputFieldVisitor

Visitor which allows a backend to visit the input fields of a Resource.

ManyToManyRelation

A many-to-many relation.

ManyToOneRelation

A many-to-one relation.

Relation

Metadata about a relation of a resource.

RelationPredicate

A boolean predicate on a Relation targeting items of type T.

RelationPredicateCompiler

A generic interface to a backend-specific relation predicate compiler.

RelationVisitor

Visitor which allows relations of a Resource to describe themselves to a backend.

Resource

A complex type in the relational GraphQL type system.

ResourceBuilder

A backend specific interface to query results, used to reconstruct a Resource.

ResourceInput

The input used to specify a new object of type T.

ResourcePredicate

A boolean predicate on a resource type T.

ResourceVisitor

Visitor which allows a Resource to describe itself to a backend.

Derive Macros

Resource

Derive an implementation of Resource, and related items, for a struct.