Derive Macro relational_graphql::graphql::type_system::resource::Resource

#[derive(Resource)]
{
    // Attributes available to this derive:
    #[resource]
}

Derive an implementation of Resource, and related items, for a struct.

This macro will derive an implementation of Resource for a struct, along with all of the necessary types to describe the resource’s predicates, fields and relations. It will also generate an async_graphql #[Object] impl block with resolvers for each of the structs fields.

Specifically, the following items are generated:

An #[Object] impl block with a resolver for each field. For regular fields, the resolver simply returns a reference to the field, which must be an OutputType. You can use the skip attribute to avoid generating the resolver for fields which are not output types. For relations, the resolver is paginated: it takes Relay-style paging arguments and loads the appropriate page of results on demand.

A unit struct for each field of the original struct, containing metadata about the field via the Field or Relation trait. These are generated in a nested module called fields.

An input type for the struct, which contains only the struct’s input fields, with references to other resources replaced by the resource ID and relations replaced by a placeholder (since they are loaded later, on demand).

A has predicate, which is a GraphQL input type allowing the client to apply a filter to any of the struct’s fields.

A predicate used to filter items of this resource. The predicate is an enum with one variant for the has predicate and, if the struct has a primary field, another variant to filter directly by the primary field.

A quantified predicate which applies to a collection of items of this resource by requiring that a certain number of items in the collection match the predicate.

A relation predicate which applies to a collection of items of this resource. The relation predicate is an enum which has variants requiring that at least or at most n items match a predicate, that any, all, or none items match a predicate, and, if the struct has a primary field, that the collection includes a given value of the primary field.

All of these items are placed in a module with the same visibility as the original struct. The items have the same visibility as the original struct, unless the original struct is private, in which case the generated items are pub(super) (so that they are visible one level up from the generated module, in the scope where the original struct was defined). The name of the generated module is, by default, the name of the struct converted to snake_case. This can be changed with the module attribute.

Documentation (doc comments or the #[doc = "..."] attribute) on the struct and its fields is automatically propagated to derived items and will appear in the exported GraphQL schema.

Examples

Derive a Resource.

use relational_graphql::prelude::*;

/// A new resource.
#[derive(Clone, Resource)]
struct MyResource {
    id: Id,
    /// A regular field.
    #[resource(primary)]
    field: u64,
    /// A relation.
    relateds: Many<Related>,
    /// A field that is not exposed to GraphQL.
    #[resource(skip)]
    extra: WeirdType,
}

/// A resource with a relation to the first resource.
#[derive(Clone, Resource)]
struct Related {
    id: Id,
    #[resource(primary)]
    name: String,
    my_resources: Many<MyResource>,
}

#[derive(Clone, Default)]
struct WeirdType;

Use it in a GraphQL schema.

use relational_graphql::{graphql::{backend::Connection, EmptyFields}};

struct Query;

#[Object]
impl Query {
    async fn my_resource(
        &self,
        #[graphql(name = "where")] filter: <MyResource as Type>::Predicate,
    ) -> MyResource {
        MyResource {
            id: 0,
            field: 0,
            // At first, we just return a placeholder for relations. If this relation is
            // requested in the GraphQL result set, we will load it on-demand in a separate
            // operation.
            relateds: Many::default(),
            extra: WeirdType,
        }
    }
}

let schema = Schema::build(Query, EmptyMutation, EmptySubscription).finish();

// Query by field.
schema
    .execute(r#"{
        myResource(
            where: {
                has: {
                    relateds: {
                        any: { is: { is: { lit: "foo" } } }
                    }
                }
            }
        ) {
            field
        }
    }"#)
    .await
    .into_result()
    .unwrap();

// Query by primary field.
schema
    .execute("{
        myResource(
            where: {
                is: { is: { lit: 0 } }
            }
        ) {
            field
        }
    }")
    .await
    .into_result()
    .unwrap();

Struct attributes

Attribute	Description	Arg	Required

plural	Override the default pluralization of the struct name. The default simply appends an s.	ident	no

module	Override the default module name for derived items. The name defaults to the snake_case version of the struct name.	ident	no

Field attributes

Attribute	Description	Arg	Required

id	Use this field as the ID for this resource. Each resource must have exactly one ID field of type Id. This attribute can be omitted if the field’s type is explicitly Id (but the macro is required if, e.g., the type of the field is a type alias).	n/a	yes

inverse	Name the field or relation on the target type of this relation which is the inverse of this relation. The argument should be an identifier naming the corresponding field on the target type. If not specified, the default name dervies from the name of this resource.	ident	no

primary	Mark this field as primary. The primary field may be used in place of the whole object in GraphQL predicates. A struct can have at most one primary field.	n/a	no

skip	Do not include this field in the GraphQL types. The type of the field must implement [Default], unless an explicit default initializer is provided (the second form).	n/a	no

skip	Skip a field, reconstructing it with the given expression when loading this object.	expr	no