Trait relational_graphql::graphql::type_system::resource::Relation

pub trait Relation: 'static {
    type Owner: Resource;
    type Target: Resource;

    const NAME: &'static str;

    // Required methods
    fn visit<V: RelationVisitor<Self::Owner>>(visitor: &mut V) -> V::Output;
    fn get_predicate(
        predicate: &<Self::Owner as Resource>::ResourcePredicate
    ) -> Option<&<Self::Target as Resource>::RelationPredicate>;
    fn get_predicate_mut(
        predicate: &mut <Self::Owner as Resource>::ResourcePredicate
    ) -> Option<&mut <Self::Target as Resource>::RelationPredicate>;
    fn take_predicate(
        predicate: &mut <Self::Owner as Resource>::ResourcePredicate
    ) -> Option<<Self::Target as Resource>::RelationPredicate>;
}

Metadata about a relation of a resource.

A relation is like a virtual property of an object. Rather than being represented explicitly in the object itself, like a Field, the value of a relation comes from the relationshp of the owning object to some other object or objects in the ontology – the target of the relation.

For example, an object of Resource A might be related to objects of Resource B by a field on B which stores the ID of type A. This would be a one-to-many relation.

We categorize relations by their arity – the number of objects on each side of the relationship. If any number of objects of one resource can be related to any number of objects of another, it is a many-to-many relationship. If an object of one resource is related to a unique object of another, but any number of the former can relate to the same object of the latter, it is a many-to-one relationship. One-to-one relations are not supported.

A useful property of a Relation is its inverse – the field or relation on the target resource which relates back to this relation. The type of the inverse depends on the arity of the relation. In a many-to-many relation, both sides are a relation, so ManyToManyRelation::Inverse implements Relation. But in a many-to-one relation, each object in the target resource (the “many” side) is related to exactly one object in the owning resource (the “one” side) by means of a field storing the ID of the owning resource, so ManyToOneRelation::Inverse implements Field.

Required Associated Types

type Owner: Resource

The resource owning this relation.

type Target: Resource

The type of objects in this relation.

Required Associated Constants

const NAME: &'static str

The name of this relation.

Required Methods

fn visit<V: RelationVisitor<Self::Owner>>(visitor: &mut V) -> V::Output

Describe this relation to a backend.

fn get_predicate( predicate: &<Self::Owner as Resource>::ResourcePredicate ) -> Option<&<Self::Target as Resource>::RelationPredicate>

Access the sub-predicate used to filter this relation in a ResourcePredicate.

fn get_predicate_mut( predicate: &mut <Self::Owner as Resource>::ResourcePredicate ) -> Option<&mut <Self::Target as Resource>::RelationPredicate>

Access the sub-predicate used to filter this relation in a ResourcePredicate.

fn take_predicate( predicate: &mut <Self::Owner as Resource>::ResourcePredicate ) -> Option<<Self::Target as Resource>::RelationPredicate>

Take the sub-predicate used to fitler this relation in a ResourcePredicate.

The next time this function, get_predicate, or get_predicate_mut is called on the same relation, it will return None, since the field has been taken.