Trait relational_graphql::graphql::type_system::resource::Resource

pub trait Resource: Type<Predicate = Self::ResourcePredicate, Input = Id> {
    type NumFields: Length;
    type NumInputFields: Length;
    type NumRelations: Length;
    type Id: Field<Resource = Self, Type = Id>;
    type ResourcePredicate: ResourcePredicate<Self>;
    type RelationPredicate: RelationPredicate<Self>;
    type ResourceInput: ResourceInput<Self>;

    // Required methods
    fn build_resource<B: ResourceBuilder<Self>>(
        builder: B
    ) -> Result<Self, B::Error>;
    fn describe_fields<V: FieldVisitor<Self>>(
        visitor: &mut V
    ) -> Array<V::Output, Self::NumFields>;
    fn describe_input_fields<V: InputFieldVisitor<Self>>(
        visitor: &mut V
    ) -> Array<V::Output, Self::NumInputFields>;
    fn describe_relations<V: RelationVisitor<Self>>(
        visitor: &mut V
    ) -> Array<V::Output, Self::NumRelations>;

    // Provided methods
    fn get<F: Field<Resource = Self>>(&self) -> &F::Type { ... }
    fn describe_resource<V: ResourceVisitor<Self>>(visitor: V) -> V::Output { ... }
    fn field_names() -> Array<&'static str, Self::NumFields> { ... }
    fn input_field_names() -> Array<&'static str, Self::NumInputFields> { ... }
}

A complex type in the relational GraphQL type system.

Required Associated Types

type NumFields: Length

The number of singular fields this resource has.

type NumInputFields: Length

The number of input fields this resource has.

type NumRelations: Length

The number of relations this resource has to other resources.

type Id: Field<Resource = Self, Type = Id>

The unique, sequentially increasing ID for this resource.

type ResourcePredicate: ResourcePredicate<Self>

Boolean predicates on this resource type.

This is always the same type as Predicate, but the alias ResourcePredicate has the more expressive trait bound ResourcePredicate instead of the generic Predicate.

type RelationPredicate: RelationPredicate<Self>

A boolean predicate on a Relation targeting items of this type.

type ResourceInput: ResourceInput<Self>

The input used to specify a new object of this type.

The input contains input values for each of the input fields of this resource.

Required Methods

fn build_resource<B: ResourceBuilder<Self>>( builder: B ) -> Result<Self, B::Error>

Build a resource using a builder supplied by the backend.

This performs the same operation as build, but it can be called directly with a ResourceBuilder, instead of the more generic Builder. This is useful when it is known that a Type is actually a Resource.

It is an invariant that for all T: Resource, T::build(builder) == T::builder_resource(builder.resource()).

fn describe_fields<V: FieldVisitor<Self>>( visitor: &mut V ) -> Array<V::Output, Self::NumFields>

Describe the fields of this resource.

fn describe_input_fields<V: InputFieldVisitor<Self>>( visitor: &mut V ) -> Array<V::Output, Self::NumInputFields>

Describe the input fields of this resource.

fn describe_relations<V: RelationVisitor<Self>>( visitor: &mut V ) -> Array<V::Output, Self::NumRelations>

Describe the relations that this resource has to other resources.

Provided Methods

fn get<F: Field<Resource = Self>>(&self) -> &F::Type

Access the field F of this Resource.

fn describe_resource<V: ResourceVisitor<Self>>(visitor: V) -> V::Output

Describe the structure and definition of this Resource.

fn field_names() -> Array<&'static str, Self::NumFields>

The names of this resource’s singular fields.

fn input_field_names() -> Array<&'static str, Self::NumInputFields>

The names of this resource’s input fields.