Trait relational_graphql::graphql::type_system::resource::ResourcePredicateCompiler

pub trait ResourcePredicateCompiler<T: Resource> {
    type Result;

    // Required methods
    fn fields(self, preds: FieldsPredicate<T>) -> Self::Result;
    fn matches(self, query: String) -> Self::Result;
    fn any<I>(self, preds: I) -> Self::Result
       where I: IntoIterator<Item = T::ResourcePredicate>;
    fn all<I>(self, preds: I) -> Self::Result
       where I: IntoIterator<Item = T::ResourcePredicate>;
}

A generic interface to a backend-specific compiler for predicates on resource.

A ResourcePredicate can use this interface to instruct an arbitrary backend on how to compile itself into a backend-specific format.

Required Associated Types

type Result

The backend-specific compilation result.

Required Methods

fn fields(self, preds: FieldsPredicate<T>) -> Self::Result

Instruct the backend to compile a predicate based on sub-predicates for each field.

The result is a predicate which is satisfied on T if the appropriate sub-predicate is satisfied on each field of T.

fn matches(self, query: String) -> Self::Result

Instruct the backend to compile a predicate based on a query string.

The result is a predicate which is satisfied if the query string matches the resource’s searchable fields. The matching algorithm is an implementation-defined, text-based matching which tries to be as powerful and general-purpose as possible. It prioritizes efficiency and efficacy over predictability, so, unlike the rest of this system, the results may depend on a complex, opaque algorithm or even a source of non-determinism.

This type of predicate provides functionality similar to general-purpose search engines.

fn any<I>(self, preds: I) -> Self::Resultwhere I: IntoIterator<Item = T::ResourcePredicate>,

Instruct the backend to compile a disjunction of predicates.

The result is a predicate which is satisfied if any of preds are.

fn all<I>(self, preds: I) -> Self::Resultwhere I: IntoIterator<Item = T::ResourcePredicate>,

Instruct the backend to compile a conjunction of predicates.

The result is a predicate which is satisfied if all of preds are.