Module relational_graphql::graphql::type_system::scalar

In GraphQL, primitives are called scalars.

Scalars make up the leaves of a graph-oriented data model. They are primitive types like integers and strings, upon which more complex object types are recursively built.

Unlike complex Resource types, scalars do not have properties or relationships to other objects. They are simple, atomic objects whose only property is their value. Users also cannot create new scalar types, so backends are able to rely on there being a known, finite set of scalar types upon which to build their own data model.

Re-exports

pub use i32_scalar::Trait as I32Scalar;

pub use i64_scalar::Trait as I64Scalar;

pub use u32_scalar::Trait as U32Scalar;

pub use u64_scalar::Trait as U64Scalar;

Modules

i32_scalar

i64_scalar

u32_scalar

u64_scalar

Structs

StringCmp

A comparison on strings.

Enums

IntCmpOp

Comparison operators for integral scalar types.

StringCmpOp

Comparison operators for strings.

StringPredicate

A boolean predicate on strings.

Value

A scalar value.

Traits

IntScalar

Integral scalars.

Scalar

A primitive type in the relational GraphQL type system.

ScalarPredicate

A boolean predicate on a scalar type T.

ScalarPredicateCompiler

A generic interface to a backend-specific compiler for predicates on scalars.

ScalarVisitor

This trait proves that a scalar is one of the supported scalar types.

StringScalar

A string scalar.

Type Definitions

Id

A type for unique, sequentially increasing IDs.