Trait relational_graphql::graphql::type_system::scalar::Scalar

pub trait Scalar: Type<Predicate = Self::ScalarPredicate, Input = Self> + InputType + ScalarType + Sealed {
    type ScalarPredicate: ScalarPredicate<Self>;
    type Cmp: Display;

    // Required method
    fn visit<V: ScalarVisitor<Self>>(visitor: V) -> V::Output;
}

A primitive type in the relational GraphQL type system.

Required Associated Types

type ScalarPredicate: ScalarPredicate<Self>

Boolean predicates on this scalar type.

This is always the same type as Predicate, but the alias ScalarPredicate has the more expressive trait bound ScalarPredicate instead of the generic Predicate.

type Cmp: Display

Comparison operators for this type of scalar.

Required Methods

fn visit<V: ScalarVisitor<Self>>(visitor: V) -> V::Output

Perform a type-level pattern match on this scalar.

This allows backend implementations to use the fact that there is a known, finite set of supported scalar types. The backend can thus use backend-specific properties of each supported type without placing backend-specific constraints on this trait itself.

The given ScalarVisitor must handle the case where Self is any of the supported scalar types.