Trait relational_graphql::graphql::type_system::scalar::ScalarPredicate

pub trait ScalarPredicate<T: Scalar>: Predicate<T> + Sealed<T> {
    // Required method
    fn compile<C: ScalarPredicateCompiler<T>>(self, compiler: C) -> C::Result;
}

A boolean predicate on a scalar type T.

Required Methods

fn compile<C: ScalarPredicateCompiler<T>>(self, compiler: C) -> C::Result

Compile this predicate into a form which the backend can execute.

When a backend data source executes a GraphQL query, it must compile each predicate in the query into a form which can be applied to data in the backend’s particular datda model. For scalar predicates, the backend implementation will construct a ScalarPredicateCompiler which is specific to that backend and pass it to compile. The predicate will use the backend-agnostic ScalarPredicateCompiler to describe the structure of this predicate and instruct the backend on how to compile it.