Trait relational_graphql::graphql::type_system::scalar::ScalarVisitor

pub trait ScalarVisitor<T: Scalar> {
    type Output;

    // Required methods
    fn visit_i32(self) -> Self::Output
       where T: I32Scalar;
    fn visit_i64(self) -> Self::Output
       where T: I64Scalar;
    fn visit_u32(self) -> Self::Output
       where T: U32Scalar;
    fn visit_u64(self) -> Self::Output
       where T: U64Scalar;
    fn visit_string(self) -> Self::Output
       where T: StringScalar;
}

This trait proves that a scalar is one of the supported scalar types.

In order to implement a backend, it is often essential to have a finite, known set of scalar types which the backend must represent. For example, an RDBMS backend will need to map scalar types to the primitive types supported by the database (e.g. integer, text, etc.). This trait proves that, given a bound T: Scalar, T is one of the following supported scalar types:

i32

i64

u32

u64

String and it proves this in a way that makes this fact usable in Rust code. This is stronger than the [macro@sealed] mechanism (which we also use for scalars) which prevents other implementations of the Scalar trait, but which does not allow the Rust compiler to use the information that there are no other implementations.

This trait allows a backend to pattern match on a Scalar type, providing functions to handle only the supported cases. This effectively lists Rust’s total pattern matching on enums to the type level. Just like in each arm of a match expression you can use the fact that the variable being matched is a particular variant of an enum, in each method of this trait you can use the fact that T is a certain scalar type, by means of the [Is] trait for type-level equality.

Examples

Suppose we want to check if a scalar value is a default value (0 for integer types and empty for strings). We could write the following function:

fn is_default_strict<T: Scalar + Default + PartialEq>(value: &T) -> bool {
    *value == T::default()
}

But this function has a strict type bound: it is only callable if T satisfies the extra trait bounds Default and PartialEq. We will have to propagate these bounds upward through the call stack, which may be difficult if, for example, this function is being called on an arbitrary field of a Resource type. We can’t easily write the constraint that all fields of the Resource we are working with implement these extra bounds. Indeed, take the ResourceBuilder::field method. It must work for all F: Field, not just F: Field, F::Type: Default + PartialEq.

To remove the extra type bounds, we can leverage the fact that a Scalar must be one of a few types, all of which happen to implement Default and PartialEq, by performing a total pattern match on the type of the scalar:

fn is_default<T: Scalar>(value: &T) -> bool {
    struct Visitor<'a, T>(&'a T);

    impl<'a, T: Scalar> ScalarVisitor<T> for Visitor<'a, T> {
        type Output = bool;

        fn visit_i32(self) -> bool
        where
            T: I32Scalar
        {
            // Here we can use the fact that `T` is `i32` by casting `self.0` from `&T` to
            // `&i32`, and then use `i32::default()`.
            let value = self.0.into_ref();
            // Now we can call the stricter typed version of the function to handle the
            // rest:
            is_default_strict(value)
        }

        // The remaining cases are similar:
        fn visit_i64(self) -> bool
        where
            T: I64Scalar
        {
            is_default_strict(self.0.into_ref())
        }
        fn visit_u32(self) -> bool
        where
            T: U32Scalar
        {
            is_default_strict(self.0.into_ref())
        }
        fn visit_u64(self) -> bool
        where
            T: U64Scalar
        {
            is_default_strict(self.0.into_ref())
        }
        fn visit_string(self) -> bool
        where
            T: StringScalar
        {
            is_default_strict(self.0.into_ref())
        }
    }

    T::visit(Visitor(value))
}

The more permissive is_default requires a lot more code than the stricter is_default_strict, because we cannot treat all cases the same using a single trait bound. Instead we must explicitly handle each case, proving to the Rust compiler that in each one, the type in question works in the way we want to use it. Still, this type of pattern matching may be a crucial function for backend implementations in case they require some specific extra trait bound on scalar types and don’t have a good place to put it.

Required Associated Types

type Output

The type of value which is returned by this type-level match.

Required Methods

fn visit_i32(self) -> Self::Outputwhere T: I32Scalar,

Handle the case where T is i32.

fn visit_i64(self) -> Self::Outputwhere T: I64Scalar,

Handle the case where T is i64.

fn visit_u32(self) -> Self::Outputwhere T: U32Scalar,

Handle the case where T is u32.

fn visit_u64(self) -> Self::Outputwhere T: U64Scalar,

Handle the case where T is u64.

fn visit_string(self) -> Self::Outputwhere T: StringScalar,

Handle the case where T is String.