Derive Macro relational_graphql::prelude::Query

#[derive(Query)]
{
    // Attributes available to this derive:
    #[query]
}

Derive resolvers for top-level query fields in a GraphQL API.

This macro generates an async_graphql #[Object] impl block for a struct, which includes resolvers for a number of given resource types. These resources act as the entrypoints to an ontology, the rest of which is implied by the transitive closure of the entrypoint resources’ fields.

The resources to include as entrypoints are given using the resource attribute, as in #[query(resource(entrypoint: ResourceType))]. entrypoint is the name of the GraphQL field which will be generated on the query object. It must be a Rust identifier (but will be converted from snake case to camel case in the GraphQL schema). ResourceType is the type of items in the collection returned by entrypoint. You may specify multiple entrypoints that resolve to the same resource type as long as they have different names.

Each entrypoint defined this way will generate a GraphQL field which takes the plural resource inputs (a where predicate and Relay paging inputs) and produces a paginated connection of items of the resource type matching the where clause.

The macro also generates a register associated function on the query type, with the signature:

pub async fn register<D: DataSource>(db: &mut D) -> Result<(), D::Error>;

This function will traverse the list of resources in the query object and register all of the reachable resource types in the database, effectively creating a relational schema for the ontology defined by this object.

Examples

use relational_graphql::prelude::*;

#[derive(Clone, Resource)]
struct MyResource {
    id: Id,
    #[resource(primary)]
    field: u64,
}

#[derive(Query)]
#[query(resource(my_resources: MyResource))]
struct Query;

let schema = Schema::build(Query, EmptyMutation, EmptySubscription).finish();

// Query by field.
schema
    .execute("{
        myResources(
            where: {
                is: { is: { lit: 0 } }
            }
        ) {
            edges { node { field } }
        }
    }")
    .await
    .into_result()
    .unwrap();

Struct attributes

Attribute	Description	Arg	Required

resource	Resource which should included in the ontology	field	yes