Module relational_graphql::sql::db

Abstract interface to a SQL database.

Modules

postgres

Instantiation of the abstract db interface for PostgreSQL.

temp

A temporary database which is dropped on Drop.

Structs

Column

An identifier of a column in a SQL query.

JoinClause

A JOIN clause.

SchemaColumn

A column in a schema.

Enums

Boolean

A boolean expression in a WHERE clause.

Clause

A clause modifying a SQL statement.

ConstraintKind

A constraint on a set of columns in a table.

FromItem

A table or view which can be referenced in the FROM clause of another query.

SelectColumn

A column in a list of columns selected from a query.

Type

A SQL primitive data type.

Value

A primitive value supported by a SQL database.

WhereClause

A WHERE clause.

Traits

AlterTable

An ALTER TABLE statement which can be executed against the database.

AlterTableExt

An extension trait for AlterTable that provides some higher-level functions.

Connection

A connection to the database.

CreateTable

A CREATE TABLE statement which can be executed against the database.

CreateTableExt

An extension trait for CreateTable that provides some higher-level functions.

Error

Errors returned by the database.

Insert

An INSERT statement which can be executed against the database.

Row

A row in a database table.

Select

A SELECT query which can be executed against the database.

SelectExt

An extension trait for Select that provides some higher-level functions.

Update

An UPDATE statement which can be executed against the database.

Functions

escape_ident

Escape an identifier (table name, column name, etc.) for inclusion in a SQL query.