Trait relational_graphql::sql::db::Connection

pub trait Connection {
    type Error: Error;
    type CreateTable<'a, N: Length>: CreateTable<Error = Self::Error>
       where Self: 'a;
    type AlterTable<'a>: AlterTable<Error = Self::Error>
       where Self: 'a;
    type Select<'a>: Select<'a, Error = Self::Error>
       where Self: 'a;
    type Insert<'a, N: Length>: Insert<N, Error = Self::Error>
       where Self: 'a;

    // Required methods
    fn create_db<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        name: &'life1 str
    ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn drop_db<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        name: &'life1 str
    ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn create_table<'a, N: Length>(
        &'a self,
        table: impl Into<Cow<'a, str>> + Send,
        columns: Array<SchemaColumn<'a>, N>
    ) -> Self::CreateTable<'a, N>;
    fn alter_table<'a>(
        &'a self,
        table: impl Into<Cow<'a, str>> + Send
    ) -> Self::AlterTable<'a>;
    fn select<'a>(
        &'a self,
        columns: &'a [SelectColumn<'a>],
        table: impl Into<Cow<'a, str>> + Send
    ) -> Self::Select<'a>;
    fn insert<'a, C, N>(
        &'a self,
        table: impl Into<Cow<'a, str>> + Send,
        columns: Array<C, N>
    ) -> Self::Insert<'a, N>
       where C: Into<String>,
             N: Length;
}

A connection to the database.

Required Associated Types

type Error: Error

Errors returned from queries.

type CreateTable<'a, N: Length>: CreateTable<Error = Self::Error> where Self: 'a

A CREATE TABLE statement which can be executed against the database.

type AlterTable<'a>: AlterTable<Error = Self::Error> where Self: 'a

An ALTER TABLE statement which can be executed against the database.

type Select<'a>: Select<'a, Error = Self::Error> where Self: 'a

A SELECT query which can be executed against the database.

type Insert<'a, N: Length>: Insert<N, Error = Self::Error> where Self: 'a

An INSERT statement which can be executed against the database.

Required Methods

fn create_db<'life0, 'life1, 'async_trait>( &'life0 mut self, name: &'life1 str ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Create a new database and connect to it.

fn drop_db<'life0, 'life1, 'async_trait>( &'life0 mut self, name: &'life1 str ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Drop the named database.

fn create_table<'a, N: Length>( &'a self, table: impl Into<Cow<'a, str>> + Send, columns: Array<SchemaColumn<'a>, N> ) -> Self::CreateTable<'a, N>

Start a CREATE TABLE statement.

table and columns describe the name and the basic structure of the table. More fine-grained control over the table (such as adding constraints) is available via the methods on the CreateTable object.

fn alter_table<'a>( &'a self, table: impl Into<Cow<'a, str>> + Send ) -> Self::AlterTable<'a>

Start an ALTER TABLE statement.

The statement will affect table. Actions to perform on the table can be specified using the methods on the AlterTable object before executing the statement.

fn select<'a>( &'a self, columns: &'a [SelectColumn<'a>], table: impl Into<Cow<'a, str>> + Send ) -> Self::Select<'a>

Start a SELECT query.

columns indicates the columns to include in the query results. The resulting Select represents a statement of the form SELECT columns FROM table. The query can be refined, for example by adding a WHERE clause, using the approriate methods on the Select object before running it.

fn insert<'a, C, N>( &'a self, table: impl Into<Cow<'a, str>> + Send, columns: Array<C, N> ) -> Self::Insert<'a, N>where C: Into<String>, N: Length,

Start an INSERT query.

table indicates the table to insert into and columns the names of the columns in that table into which values should be inserted.